落地页工单制作/电商创意图模仿工单制作/requirements.txt
google-generativeai>=0.8.0
python-docx>=1.1.0
Pillow>=10.0.0
python-dotenv>=1.0.0
